feat(utils): add filename base ID extraction
